[REF] move action_menus in views/
